requirements.txt
Django~=2.0.9
django-userforeignkey~=0.3.0
django-simple-history~=2.6.0
djangorestframework~=3.9.0
Markdown~=3.0.1
django_filter~=2.0.0
mysqlclient~=1.3.12
pytz~=2018.7